refactor(dot): split of consolidation functions
and tweak the dir output to look a bit better (thinner edges, folders shown as folders, show a separate node for composites
