+++ src/report/dot/folderLevel/ddot.template.hbs
@@ -9,12 +9,12 @@ digraph "dependency-cruiser output"{
     fontsize=9
     style="rounded,bold"
     compound=true
-    node [shape=box style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
-    edge [color="#00000077" penwidth=2.0 arrowhead=normal fontname=Helvetica fontsize=9]
+    node [shape=folder style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
+    edge [color=black arrowhead=normal fontname=Helvetica fontsize=9]
 
 {{#each things}}
     {{#if folder}}
-    {{#each path}}subgraph "cluster_{{{aggregateSnippet}}}" {label="{{{snippet}}}" "{{{aggregateSnippet}}}" [label="" shape=point style=invis width="0"] {{/each~}}
+    {{#each path}}subgraph "cluster_{{{aggregateSnippet}}}" {label="{{{snippet}}}" "{{{aggregateSnippet}}}" [label="." width="0.5"{{!--shape=point style=invis width="0"--}}] {{/each~}}
         "{{{source}}}" [label="{{{label}}}" {{#if color}}color="{{{color}}}" {{/if}}{{#if fontcolor}}fontcolor="{{{fontcolor}}}" {{/if}}{{#if fillcolor}}fillcolor="{{{fillcolor}}}" {{/if}}{{#if matchesDoNotFollow}}shape="folder" {{/if}}{{#if rule}}tooltip="{{{rule.name}}}" {{/if}}{{~#if url}}URL="{{{url}}}"{{/if~}}]{{#each path}} }{{/each}}
     {{else}}
     "{{{source}}}" [label="{{{label}}}" {{#if color}}color="{{{color}}}" {{/if}}{{#if fontcolor}}fontcolor="{{{fontcolor}}}" {{/if}}{{#if fillcolor}}fillcolor="{{{fillcolor}}}" {{/if}}{{#if matchesDoNotFollow}}shape="folder" {{/if}}{{#if rule}}tooltip="{{{rule.name}}}" {{/if}}{{~#if url}}URL="{{{url}}}"{{/if~}}]
