+++ test/report/dot/common/dotColoring.spec.js
