+++ test/report/dot/folderLevel/fixtures/consolidated-orphans.dot
@@ -9,8 +9,8 @@ digraph "dependency-cruiser output"{
     fontsize=9
     style="rounded,bold"
     compound=true
-    node [shape=box style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
-    edge [color="#00000077" penwidth=2.0 arrowhead=normal fontname=Helvetica fontsize=9]
+    node [shape=folder style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
+    edge [color=black arrowhead=normal fontname=Helvetica fontsize=9]
 
     "tmp" [label="tmp" tooltip="no-orphans" ]
